Ronniel1

+++ src/Components/Demo.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 
-import { copy, linkIcon, loader } from "../assets";
+import { copy, linkIcon, loader, tick } from "../assets";
 import { useLazyGetSummaryQuery } from "../services/article";
 
 const Demo = () => {
@@ -8,11 +8,13 @@ const Demo = () => {
     url: "",
     summary: ""
   });
-
   const [allArticles, setAllArticles] = useState([]);
+  const [copied, setCopied] = useState("");
 
+  // RTK lazy query
   const [getSummary, { error, isFetching }] = useLazyGetSummaryQuery();
 
+  // Load data from localStorage on mount
   useEffect(() => {
     const articlesFromLocalStorage = JSON.parse(
       localStorage.getItem("articles")
@@ -26,24 +28,40 @@ const Demo = () => {
   const handleSubmit = async (e) => {
     e.preventDefault();
 
-    const { data } = await getSummary({
-      articleUrl: article.url
-    });
+    const existingArticle = allArticles.find(
+      (item) => item.url === article.url
+    );
 
+    if (existingArticle) return setArticle(existingArticle);
+
+    const { data } = await getSummary({ articleUrl: article.url });
     if (data?.summary) {
       const newArticle = { ...article, summary: data.summary };
-
       const updatedAllArticles = [newArticle, ...allArticles];
 
+      // update state and local storage
       setArticle(newArticle);
       setAllArticles(updatedAllArticles);
-
       localStorage.setItem("articles", JSON.stringify(updatedAllArticles));
+    }
+  };
+
+  // copy the url and toggle the icon for user feedback
+  const handleCopy = (copyUrl) => {
+    setCopied(copyUrl);
+    navigator.clipboard.writeText(copyUrl);
+    setTimeout(() => setCopied(false), 3000);
+  };
+
+  const handleKeyDown = (e) => {
+    if (e.keyCode === 13) {
+      handleSubmit(e);
     }
   };
 
   return (
     <section className="mt-16 w-full max-w-xl">
+      {/* Search */}
       <div className="flex flex-col w-full gap-2">
         <form
           className="relative flex justify-center items-center"
@@ -51,39 +69,39 @@ const Demo = () => {
         >
           <img
             src={linkIcon}
-            alt="link_icon"
+            alt="link-icon"
             className="absolute left-0 my-2 ml-3 w-5"
           />
 
           <input
             type="url"
-            placeholder="Place a url"
+            placeholder="Paste the article link"
             value={article.url}
             onChange={(e) => setArticle({ ...article, url: e.target.value })}
+            onKeyDown={handleKeyDown}
             required
-            className="url_input peer"
+            className="url_input peer" // When you need to style an element based on the state of a sibling element, mark the sibling with the peer class, and use peer-* modifiers to style the target element
           />
-
           <button
             type="submit"
             className="submit_btn peer-focus:border-gray-700 peer-focus:text-gray-700 "
           >
-            ↵
+            <p>↵</p>
           </button>
         </form>
 
-        {/* Browser URL history */}
+        {/* Browse History */}
         <div className="flex flex-col gap-1 max-h-60 overflow-y-auto">
-          {allArticles.map((item, index) => (
+          {allArticles.reverse().map((item, index) => (
             <div
               key={`link-${index}`}
-              onClick={() => setAllArticles(item)}
+              onClick={() => setArticle(item)}
               className="link_card"
             >
-              <div className="copy_btn">
+              <div className="copy_btn" onClick={() => handleCopy(item.url)}>
                 <img
-                  src={copy}
-                  alt="copy_icon"
+                  src={copied === item.url ? tick : copy}
+                  alt={copied === item.url ? "tick_icon" : "copy_icon"}
                   className="w-[40%] h-[40%] object-contain"
                 />
               </div>
@@ -95,13 +113,13 @@ const Demo = () => {
         </div>
       </div>
 
-      {/* Display results  */}
+      {/* Display Result */}
       <div className="my-10 max-w-full flex justify-center items-center">
         {isFetching ? (
           <img src={loader} alt="loader" className="w-20 h-20 object-contain" />
         ) : error ? (
           <p className="font-inter font-bold text-black text-center">
-            Well, that wasn&#39t supposed to happen
+            Well, that wasn't supposed to happen...
             <br />
             <span className="font-satoshi font-normal text-gray-700">
               {error?.data?.error}
@@ -125,4 +143,5 @@ const Demo = () => {
     </section>
   );
 };
+
 export default Demo;
